+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `metrics.cfg{}` `"all"` metasection for array `include` and `exclude`
+  (`metrics.cfg{include={'all'}}` can be used instead of `metrics.cfg{include='all'}`,
+  `metrics.cfg{exclude={'all'}}` can be used instead of `metrics.cfg{include='none'}`)
+
 ## [1.0.0] - 2023-05-22
 ### Changed